requirements.txt
pytest==8.4.2
playwright == 1.55.0
requests==2.32.5
pip == 25.1.1
pyee == 13.0.0
Pygments ==2.19.2
pluggy == 1.6.0
iniconfig == 2.1.0
pymongo == 4.15.3
allure-pytest == 2.15.0
allure-python-commons == 2.15.0
jsonschema == 4.25.1
jsonschema-specifications == 2025.9.1